Use explicily 64-bit for RAM and Disk sizes
While we will only use 64-bit machines, and on these it will be 64-bit
by default, it doesn't hurt to be explicit about it.

+++ api/v1/eviction_types.go
@@ -49,7 +49,7 @@ type EvictionStatus struct {
 	HypervisorServiceId string `json:"hypervisorServiceId"`
 	// +kubebuilder:validation:Optional
 	// +kubebuilder:default=0
-	OutstandingRamMb int `json:"outstandingRamMb"`
+	OutstandingRamMb int64 `json:"outstandingRamMb"`
 	// +kubebuilder:validation:Optional
 	OutstandingInstances []string `json:"outstandingInstances"`
